refactor(utils): restructure number utils
