+++ packages/lumx-react/src/components/slider/index.ts
@@ -1,2 +1,2 @@
 export * from './Slider';
-export { clamp } from '@lumx/react/utils/clamp';
+export { clamp } from '@lumx/react/utils/number/clamp';
