+++ packages/lumx-react/src/utils/number/clamp.ts
